change the access modifier for sample code to hide APIs from documents
