+++ RenderStreaming~/ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2020.3.27f1
-m_EditorVersionWithRevision: 2020.3.27f1 (e759542391ea)
+m_EditorVersion: 2020.3.28f1
+m_EditorVersionWithRevision: 2020.3.28f1 (f5400f52e03f)
