+++ com.unity.renderstreaming/Samples~/Example/Scripts/UIController.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 using UnityEngine.UI;
 using UnityEngine.InputSystem;
 
-namespace Unity.RenderStreaming
+namespace Unity.RenderStreaming.Samples
 {
     [RequireComponent(typeof(RectTransform))]
     public class UIController : MonoBehaviour
